processor/owner(ticdc): fix potential data race in ddl sink
