+++ cdc/owner/changefeed_test.go
@@ -545,243 +545,6 @@ func testChangefeedReleaseResource(
 	require.True(t, os.IsNotExist(err))
 }
 
-func TestAddSpecialComment(t *testing.T) {
-	testCase := []struct {
-		input  string
-		result string
-	}{
-		{
-			"create table t1 (id int ) shard_row_id_bits=2;",
-			"CREATE TABLE `t1` (`id` INT) /*T! SHARD_ROW_ID_BITS = 2 */",
-		},
-		{
-			"create table t1 (id int ) shard_row_id_bits=2 pre_split_regions=2;",
-			"CREATE TABLE `t1` (`id` INT) " +
-				"/*T! SHARD_ROW_ID_BITS = 2 */ /*T! PRE_SPLIT_REGIONS = 2 */",
-		},
-		{
-			"create table t1 (id int ) shard_row_id_bits=2     pre_split_regions=2;",
-			"CREATE TABLE `t1` (`id` INT) " +
-				"/*T! SHARD_ROW_ID_BITS = 2 */ /*T! PRE_SPLIT_REGIONS = 2 */",
-		},
-		{
-			"create table t1 (id int ) shard_row_id_bits=2 engine=innodb pre_split_regions=2;",
-			"CREATE TABLE `t1` (`id` INT) /*T! SHARD_ROW_ID_BITS = 2 */" +
-				" ENGINE = innodb /*T! PRE_SPLIT_REGIONS = 2 */",
-		},
-		{
-			"create table t1 (id int ) pre_split_regions=2 shard_row_id_bits=2;",
-			"CREATE TABLE `t1` (`id` INT) /*T! PRE_SPLIT_REGIONS = 2 */" +
-				" /*T! SHARD_ROW_ID_BITS = 2 */",
-		},
-		{
-			"create table t6 (id int ) " +
-				"shard_row_id_bits=2 shard_row_id_bits=3 pre_split_regions=2;",
-			"CREATE TABLE `t6` (`id` INT) /*T! SHARD_ROW_ID_BITS = 2 */ " +
-				"/*T! SHARD_ROW_ID_BITS = 3 */ /*T! PRE_SPLIT_REGIONS = 2 */",
-		},
-		{
-			"create table t1 (id int primary key auto_random(2));",
-			"CREATE TABLE `t1` (`id` INT PRIMARY KEY /*T![auto_rand] AUTO_RANDOM(2) */)",
-		},
-		{
-			"create table t1 (id int primary key auto_random);",
-			"CREATE TABLE `t1` (`id` INT PRIMARY KEY /*T![auto_rand] AUTO_RANDOM */)",
-		},
-		{
-			"create table t1 (id int auto_random ( 4 ) primary key);",
-			"CREATE TABLE `t1` (`id` INT /*T![auto_rand] AUTO_RANDOM(4) */ PRIMARY KEY)",
-		},
-		{
-			"create table t1 (id int  auto_random  (   4    ) primary key);",
-			"CREATE TABLE `t1` (`id` INT /*T![auto_rand] AUTO_RANDOM(4) */ PRIMARY KEY)",
-		},
-		{
-			"create table t1 (id int auto_random ( 3 ) primary key) auto_random_base = 100;",
-			"CREATE TABLE `t1` (`id` INT /*T![auto_rand] AUTO_RANDOM(3) */" +
-				" PRIMARY KEY) /*T![auto_rand_base] AUTO_RANDOM_BASE = 100 */",
-		},
-		{
-			"create table t1 (id int auto_random primary key) auto_random_base = 50;",
-			"CREATE TABLE `t1` (`id` INT /*T![auto_rand] AUTO_RANDOM */ PRIMARY KEY)" +
-				" /*T![auto_rand_base] AUTO_RANDOM_BASE = 50 */",
-		},
-		{
-			"create table t1 (id int auto_increment key) auto_id_cache 100;",
-			"CREATE TABLE `t1` (`id` INT AUTO_INCREMENT PRIMARY KEY) " +
-				"/*T![auto_id_cache] AUTO_ID_CACHE = 100 */",
-		},
-		{
-			"create table t1 (id int auto_increment unique) auto_id_cache 10;",
-			"CREATE TABLE `t1` (`id` INT AUTO_INCREMENT UNIQUE KEY) " +
-				"/*T![auto_id_cache] AUTO_ID_CACHE = 10 */",
-		},
-		{
-			"create table t1 (id int) auto_id_cache = 5;",
-			"CREATE TABLE `t1` (`id` INT) /*T![auto_id_cache] AUTO_ID_CACHE = 5 */",
-		},
-		{
-			"create table t1 (id int) auto_id_cache=5;",
-			"CREATE TABLE `t1` (`id` INT) /*T![auto_id_cache] AUTO_ID_CACHE = 5 */",
-		},
-		{
-			"create table t1 (id int) /*T![auto_id_cache] auto_id_cache=5 */ ;",
-			"CREATE TABLE `t1` (`id` INT) /*T![auto_id_cache] AUTO_ID_CACHE = 5 */",
-		},
-		{
-			"create table t1 (id int, a varchar(255), primary key (a, b) clustered);",
-			"CREATE TABLE `t1` (`id` INT,`a` VARCHAR(255),PRIMARY KEY(`a`, `b`)" +
-				" /*T![clustered_index] CLUSTERED */)",
-		},
-		{
-			"create table t1(id int, v int, primary key(a) clustered);",
-			"CREATE TABLE `t1` (`id` INT,`v` INT,PRIMARY KEY(`a`) " +
-				"/*T![clustered_index] CLUSTERED */)",
-		},
-		{
-			"create table t1(id int primary key clustered, v int);",
-			"CREATE TABLE `t1` (`id` INT PRIMARY KEY " +
-				"/*T![clustered_index] CLUSTERED */,`v` INT)",
-		},
-		{
-			"alter table t add primary key(a) clustered;",
-			"ALTER TABLE `t` ADD PRIMARY KEY(`a`) /*T![clustered_index] CLUSTERED */",
-		},
-		{
-			"create table t1 (id int, a varchar(255), primary key (a, b) nonclustered);",
-			"CREATE TABLE `t1` (`id` INT,`a` VARCHAR(255),PRIMARY KEY(`a`, `b`)" +
-				" /*T![clustered_index] NONCLUSTERED */)",
-		},
-		{
-			"create table t1 (id int, a varchar(255), primary key (a, b) " +
-				"/*T![clustered_index] nonclustered */);",
-			"CREATE TABLE `t1` (`id` INT,`a` VARCHAR(255),PRIMARY KEY(`a`, `b`)" +
-				" /*T![clustered_index] NONCLUSTERED */)",
-		},
-		{
-			"create table clustered_test(id int)",
-			"CREATE TABLE `clustered_test` (`id` INT)",
-		},
-		{
-			"create database clustered_test",
-			"CREATE DATABASE `clustered_test`",
-		},
-		{
-			"create database clustered",
-			"CREATE DATABASE `clustered`",
-		},
-		{
-			"create table clustered (id int)",
-			"CREATE TABLE `clustered` (`id` INT)",
-		},
-		{
-			"create table t1 (id int, a varchar(255) key clustered);",
-			"CREATE TABLE `t1` (" +
-				"`id` INT,`a` VARCHAR(255) PRIMARY KEY /*T![clustered_index] CLUSTERED */)",
-		},
-		{
-			"alter table t force auto_increment = 12;",
-			"ALTER TABLE `t` /*T![force_inc] FORCE */ AUTO_INCREMENT = 12",
-		},
-		{
-			"alter table t force, auto_increment = 12;",
-			"ALTER TABLE `t` FORCE /* AlterTableForce is not supported */ , AUTO_INCREMENT = 12",
-		},
-		{
-			"create table cdc_test (id varchar(10) primary key ,c1 varchar(10)) " +
-				"ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin" +
-				"/*!90000  SHARD_ROW_ID_BITS=4 PRE_SPLIT_REGIONS=3 */",
-			"CREATE TABLE `cdc_test` (`id` VARCHAR(10) PRIMARY KEY,`c1` VARCHAR(10)) " +
-				"ENGINE = InnoDB DEFAULT CHARACTER SET = UTF8MB4 DEFAULT COLLATE = UTF8MB4_BIN " +
-				"/*T! SHARD_ROW_ID_BITS = 4 */ /*T! PRE_SPLIT_REGIONS = 3 */",
-		},
-		{
-			"CREATE TABLE t1 (id BIGINT NOT NULL PRIMARY KEY auto_increment, " +
-				"b varchar(255)) PLACEMENT POLICY=placement1;",
-			"CREATE TABLE `t1` (`id` BIGINT NOT NULL PRIMARY KEY AUTO_INCREMENT,`b` VARCHAR(255)) ",
-		},
-		{
-			"CREATE TABLE `t1` (\n  `a` int(11) DEFAULT NULL\n) " +
-				"ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin " +
-				"/*T![placement] PLACEMENT POLICY=`p2` */",
-			"CREATE TABLE `t1` (`a` INT(11) DEFAULT NULL) " +
-				"ENGINE = InnoDB DEFAULT CHARACTER SET = UTF8MB4 DEFAULT COLLATE = UTF8MB4_BIN ",
-		},
-		{
-			"CREATE TABLE t4 (" +
-				"firstname VARCHAR(25) NOT NULL," +
-				"lastname VARCHAR(25) NOT NULL," +
-				"username VARCHAR(16) NOT NULL," +
-				"email VARCHAR(35)," +
-				"joined DATE NOT NULL) " +
-				"PARTITION BY RANGE( YEAR(joined) )" +
-				" (PARTITION p0 VALUES LESS THAN (1960) PLACEMENT POLICY=p1," +
-				"PARTITION p1 VALUES LESS THAN (1970),PARTITION p2 VALUES LESS THAN (1980)," +
-				"PARTITION p3 VALUES LESS THAN (1990),PARTITION p4 VALUES LESS THAN MAXVALUE);",
-			"CREATE TABLE `t4` (" +
-				"`firstname` VARCHAR(25) NOT NULL," +
-				"`lastname` VARCHAR(25) NOT NULL," +
-				"`username` VARCHAR(16) NOT NULL," +
-				"`email` VARCHAR(35)," +
-				"`joined` DATE NOT NULL) " +
-				"PARTITION BY RANGE (YEAR(`joined`)) " +
-				"(PARTITION `p0` VALUES LESS THAN (1960) ,PARTITION `p1` VALUES LESS THAN (1970)," +
-				"PARTITION `p2` VALUES LESS THAN (1980),PARTITION `p3` VALUES LESS THAN (1990)," +
-				"PARTITION `p4` VALUES LESS THAN (MAXVALUE))",
-		},
-		{
-			"ALTER TABLE t3 PLACEMENT POLICY=DEFAULT;",
-			"ALTER TABLE `t3`",
-		},
-		{
-			"ALTER TABLE t1 PLACEMENT POLICY=p10",
-			"ALTER TABLE `t1`",
-		},
-		{
-			"ALTER TABLE t1 PLACEMENT POLICY=p10, add d text(50)",
-			"ALTER TABLE `t1` ADD COLUMN `d` TEXT(50)",
-		},
-		{
-			"alter table tp PARTITION p1 placement policy p2",
-			"",
-		},
-		{
-			"alter table t add d text(50) PARTITION p1 placement policy p2",
-			"ALTER TABLE `t` ADD COLUMN `d` TEXT(50)",
-		},
-		{
-			"alter table tp set tiflash replica 1 PARTITION p1 placement policy p2",
-			"ALTER TABLE `tp` SET TIFLASH REPLICA 1",
-		},
-		{
-			"ALTER DATABASE TestResetPlacementDB PLACEMENT POLICY SET DEFAULT",
-			"",
-		},
-
-		{
-			"ALTER DATABASE TestResetPlacementDB PLACEMENT POLICY p1 charset utf8mb4",
-			"ALTER DATABASE `TestResetPlacementDB`  CHARACTER SET = utf8mb4",
-		},
-		{
-			"/*T![placement] ALTER DATABASE `db1` PLACEMENT POLICY = `p1` */",
-			"",
-		},
-		{
-			"ALTER PLACEMENT POLICY p3 PRIMARY_REGION='us-east-1' " +
-				"REGIONS='us-east-1,us-east-2,us-west-1';",
-			"",
-		},
-	}
-	for _, ca := range testCase {
-		re, err := addSpecialComment(ca.input)
-		require.Nil(t, err)
-		require.Equal(t, re, ca.result)
-	}
-	require.Panics(t, func() {
-		_, _ = addSpecialComment("alter table t force, auto_increment = 12;alter table t force, auto_increment = 12;")
-	}, "invalid ddlQuery statement size")
-}
-
 func TestExecRenameTablesDDL(t *testing.T) {
 	helper := entry.NewSchemaTestHelper(t)
 	defer helper.Close()
